Split out Vector and Matrix classes to their own files
